Writing pending characterization specs for the actual parse tree computation, conversion to S-expressions and process() call.
Note: we now run into a delimma, caused by the original author's failure to
write the system test-first. Basically, the initialize() method instantiates a
ParseTree instance and holds onto it in an instance variable with no accessor
method(s). This means that not only is it difficult to characterize
or put expectations on the ParseTree instance itself later, i.e., now.

I'll do a refactoring now that adds an accessor, then move as carefully as
possible to using the accessor, and then continue characterization.